+++ yuno-flutter-qa-app/lib/features/home/presenter/screen/home.dart
@@ -432,6 +432,7 @@ class _FontListNameState extends ConsumerState<FontListName> {
 }
 
 List<String> fontNamesIOS = [
+  "Dancing Script",
   "Academy Engraved LET",
   "Al Nile",
   "American Typewriter",